SimpleTwitterAnalyzer.jl

SimpleTwitterAnalyzer.jl

Documentation for SimpleTwitterAnalyzer.jl

Analyzer

SimpleTwitterAnalyzer.cntoccurences — Function.

cntoccurences(df::AbstractDataFrame, col)

Count the occurences of different values of the column col in the dataframe df. Return a new dataframe with a column col representing an instance and a column cnt indicating its occurences.

Suitable for :lang and :source tweet attributes.

source

SimpleTwitterAnalyzer.cntentities — Function.

cntentities(df::AbstractDataFrame)

Count the occurences of hashtags, mentions and domains in the dataframe df and store the computed frequencies into three different dataframes. Return a dict where each entry is associated with the corresponding frequency dataframe.

source

SimpleTwitterAnalyzer.cntactivities — Function.

cntactivities(df::AbstractDataFrame)

Count the number of tweets per week and per hour in the dataframe df. Store the computed frequencies into two different dataframes. Return a dict where each entry is associated with the corresponding frequency dataframe.

source

I/O

SimpleTwitterAnalyzer.loadtweets — Function.

loadtweets(fname::AbstractString, ntweets::Int = typemax(Int))

Load ntweets tweets from a file fname. Return a DataFrame where each column represents a tweet field.

source

loadtweets(io::IO, ntweets::Int = typemax(Int))

Load ntweets tweets from a stream io. Return a DataFrame where each column represents a tweet field.

source